final/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
pytest==6.2.4
selenium==3.141.0
pytest-rerunfailures==10.0